Update to build on barebones EL8
Several lua packages are in the powertools repository so enablerepo
has been added to the yum install for redhat.  This was discovered
when testing on a barebones RockyLinux-8 Docker container for
testing.

Also on the barebones system make is not installed by default when
building.

Changes have been tested against Centos7 and RockyLinux8.  Centos6
docker container in molecule does not work.